Add testing support for artifacts, and use it
